Moving the empty BUILD method back.
The modifier should live in the component, for sure.
However, if multiple components want to do this trick, then they'll
all need to provide empty BUILD methods (which will cause role conflict
detection to fire if they are composed together). As such, I think it's
likely to be less surprising for the user to have the empty method
here - as it makes multiple component composition (with conflict
detection) work more like applying roles one at a time.